Partie 2 - Docker-map-reduce/requirement.txt
opendatasets
Pillow
numpy 
scipy 
matplotlib 
pandas 
scikit-learn
webcolors

